requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
greenlet==3.0.3
idna==3.6
pycparser==2.21
pydantic==2.6.1
pydantic-core==2.16.2
PyMySQL==1.1.0
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.27
sqlmodel==0.0.16
typing-extensions==4.9.0
urllib3==2.2.1
